Break out MapLayers component
